% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-frame.R
\name{vec_map_dfr}
\alias{vec_map_dfr}
\alias{vec_map_dfc}
\title{Apply a function to each element of a vector and combine into a data frame}
\usage{
vec_map_dfr(.x, .f, ...)

vec_map_dfc(.x, .f, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function, formula, or vector (not necessarily atomic).

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it is
converted to an extractor function. Character vectors index by
name and numeric vectors index by position; use a list to index
by position and name at different levels. If a component is not
present, the value of \code{.default} will be returned.}

\item{...}{Additional arguments passed on to the mapped function.}
}
\description{
Unlike \code{map_dfr()}, \code{vec_map_dfr()} treats vectors as 1 row data frames.
}
\examples{

library(purrr)

x <- stats::setNames(1:5, 1:5)

map_dfr(x, ~ .x)

vec_map_dfr(x, ~ .x)

}
